+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemSyncableOutputIntegrationTest.java
@@ -16,7 +16,10 @@ package com.google.cloud.hadoop.fs.gcs;
 
 import static com.google.cloud.hadoop.fs.gcs.GoogleHadoopFileSystemConfiguration.GCS_OUTPUT_STREAM_TYPE;
 import static com.google.common.truth.Truth.assertThat;
+import static java.nio.charset.StandardCharsets.UTF_8;
+import static org.junit.Assert.assertThrows;
 
+import com.google.api.client.googleapis.json.GoogleJsonResponseException;
 import com.google.cloud.hadoop.fs.gcs.GoogleHadoopFileSystemBase.OutputStreamType;
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorageFileSystemIntegrationTest;
 import java.net.URI;
@@ -96,5 +99,53 @@ public class GoogleHadoopFileSystemSyncableOutputIntegrationTest
     }
 
     assertThat(actual).isEqualTo(expected);
+  }
+
+  @Test
+  public void append_shouldAppendNewData() throws Exception {
+    URI path = GoogleCloudStorageFileSystemIntegrationTest.getTempFilePath();
+    Path hadoopPath = ghfsHelper.castAsHadoopPath(path);
+
+    ghfsHelper.writeTextFile(path.getAuthority(), path.getPath(), "original-content");
+
+    // Test appending three 9-character strings to existing object using 20 bytes buffer size
+    try (FSDataOutputStream os = ghfs.append(hadoopPath, 20, /* progress= */ () -> {})) {
+      os.write("_append-1".getBytes(UTF_8));
+
+      // Validate that file content didn't change after write call
+      assertThat(ghfsHelper.readTextFile(hadoopPath)).isEqualTo("original-content");
+
+      os.hsync();
+
+      // Validate that hsync persisted data
+      assertThat(ghfsHelper.readTextFile(hadoopPath)).isEqualTo("original-content_append-1");
+
+      os.write("_append-2".getBytes(UTF_8));
+      os.write("_append-3".getBytes(UTF_8));
+    }
+
+    String expectedContent = "original-content_append-1_append-2_append-3";
+
+    assertThat(ghfsHelper.readTextFile(hadoopPath)).isEqualTo(expectedContent);
+
+    // Check if file after appending has right size
+    assertThat(ghfs.getFileStatus(hadoopPath).getLen()).isEqualTo(expectedContent.length());
+  }
+
+  @Test
+  public void append_shouldFail_whenFileDoesNotExist() throws Exception {
+    URI path = GoogleCloudStorageFileSystemIntegrationTest.getTempFilePath();
+    Path hadoopPath = ghfsHelper.castAsHadoopPath(path);
+
+    // Test appending three 9-character strings to existing object using 20 bytes buffer size
+    FSDataOutputStream fsos = ghfs.append(hadoopPath, 20, /* progress= */ () -> {});
+    fsos.write("_append-1".getBytes(UTF_8));
+
+    assertThrows(GoogleJsonResponseException.class, fsos::hsync);
+
+    assertThrows(NullPointerException.class, fsos::close);
+
+    // Validate that file wasn't created
+    assertThat(ghfs.exists(hadoopPath)).isFalse();
   }
 }