+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopSyncableOutputStreamTest.java
@@ -125,7 +125,8 @@ public class GoogleHadoopSyncableOutputStreamTest {
             ghfs.getGcsPath(objectPath),
             new FileSystem.Statistics(ghfs.getScheme()),
             CreateFileOptions.DEFAULT,
-            mockExecutorService);
+            mockExecutorService,
+            /* appendMode= */ false);
 
     IOException fakeIoException = new IOException("fake io exception");
     when(mockExecutorService.submit(any(Callable.class)))