+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/HadoopFileSystemTestBase.java
@@ -373,21 +373,19 @@ public abstract class HadoopFileSystemTestBase extends GoogleCloudStorageFileSys
         IOException.class, () -> ghfsHelper.writeFile(hadoopPath, text, 1, /* overwrite= */ false));
   }
 
-  /**
-   * Validates append().
-   */
+  /** Validates append(). */
   @Test
-  public void testAppend()
-      throws IOException {
+  public void testAppend() throws IOException {
     URI path = GoogleCloudStorageFileSystemIntegrationTest.getTempFilePath();
     Path hadoopPath = ghfsHelper.castAsHadoopPath(path);
-    assertThrows(
-        IOException.class,
-        () ->
-            ghfs.append(
-                hadoopPath,
-                GoogleHadoopFileSystemConfiguration.GCS_OUTPUT_STREAM_BUFFER_SIZE.getDefault(),
-                null));
+
+    ghfsHelper.writeTextFile(path.getAuthority(), path.getPath(), "content");
+
+    try (FSDataOutputStream fsos = ghfs.append(hadoopPath)) {
+      fsos.write("_appended".getBytes(UTF_8));
+    }
+
+    assertThat(ghfsHelper.readTextFile(hadoopPath)).isEqualTo("content_appended");
   }
 
   /**