+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/WebHdfsIntegrationTest.java
@@ -142,23 +142,6 @@ public class WebHdfsIntegrationTest extends HadoopFileSystemTestBase {
   // -----------------------------------------------------------------
 
   /**
-   * Validates append().
-   */
-  @Test @Override
-  public void testAppend()
-      throws IOException {
-    URI path = GoogleCloudStorageFileSystemIntegrationTest.getTempFilePath();
-    Path hadoopPath = ghfsHelper.castAsHadoopPath(path);
-    // For now, verify that append does not throw. We are not interested in
-    // verifying that append() actually appends correctly. We will do that
-    // once GHFS also starts supporting appends.
-    ghfs.append(
-        hadoopPath,
-        GoogleHadoopFileSystemConfiguration.GCS_OUTPUT_STREAM_BUFFER_SIZE.getDefault(),
-        null);
-  }
-
-  /**
    * Validates getDefaultReplication().
    */
   @Test @Override